pyproject.toml
[project]
name = "waywarden"
version = "0.1.0"
description = "Slim, profile-driven agent harness"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
  "fastapi==0.135.3",
  "uvicorn[standard]==0.44.0",
  "pydantic==2.13.0",
  "pydantic-settings==2.11.0",
  "sqlalchemy==2.0.49",
  "alembic==1.18.4",
  "psycopg[binary]==3.2.12",
  "pyyaml==6.0.3",
  "greenlet>=3.4.0",
]

[project.optional-dependencies]
anthropic = [
  "anthropic>=0.74.0",
]
honcho = [
  "honcho>=1.0.0",
]
dev = [
  "pre-commit==4.5.1",
  "pytest==9.0.3",
  "pytest-asyncio==1.3.0",
  "pytest-cov==7.1.0",
  "ruff==0.15.10",
  "mypy==1.20.1",
  "types-PyYAML==6.0.12.20260408",
  "httpx==0.28.1",
]
otel = [
  "opentelemetry-api>=1.36.0",
  "opentelemetry-sdk>=1.36.0",
]

[project.scripts]
waywarden = "waywarden.cli.main:run"

[tool.uv]
package = true
required-version = ">=0.11.0"

[tool.hatch.build.targets.wheel]
packages = ["src/waywarden"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B005", "E402", "SIM117"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src/waywarden --cov=alembic --cov-report=term-missing --cov-report=xml --cov-fail-under=80"
markers = [
  "integration: marks integration tests that require external services",
]

[tool.coverage.run]
branch = true
source = ["src/waywarden", "alembic", "tests/integration/ea"]
omit = [
  "src/waywarden/__init__.py",
  "src/waywarden/infra/tracing/otel.py",
  "alembic/versions/*.py",
]

[tool.coverage.report]
show_missing = true
exclude_also = [
  "if __name__ == .__main__.:",
  "raise NotImplementedError",
]

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "waywarden.infra.tracing.otel"
disable_error_code = "import-not-found"

[[tool.mypy.overrides]]
module = "waywarden.policy.*"
disable_error_code = [
    "import-untyped",
    "typeddict-unknown-key",
    "type-arg",
]

[[tool.mypy.overrides]]
module = "waywarden.*"
strict = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
]
